requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
python-barcode==0.15.1
Werkzeug==3.0.1
zipp==3.17.0
